+++ lumenairy/__init__.py
@@ -885,7 +885,7 @@ load_zemax_prescription_txt = _deprecated_alias(
     version_removed='5.0',
 )
 
-__version__ = "5.7.0"
+__version__ = "5.7.1"
 
 #
 # __all__ is grouped by user-journey tier: